refactor geostore registering into a helper file
